Issue  #16: Fix display of long names
Create a `UserDisplayData` which a single flat data class containing all the attributes to display for a user. It contains a single `name` field instead of first and last names.

Provide a mapping, between the api model and this ui model.

With the two textviews merged, the text now wraps and the whole name is displayed